requirements.txt
aiofiles==23.2.1
#aiogram==3.4.1
aiohttp==3.9.3
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2024.2.2
charset-normalizer==3.3.2
fake-useragent==1.4.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
psycopg2==2.9.9
pycountry==22.3.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-socks==2.4.4
pytz==2024.1
requests==2.31.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.25
syncpg==1.1.2
typing_extensions==4.9.0
urllib3==2.2.0
useragent==0.1.1
yarl==1.9.4
timeout-decorator==0.5.0
aiogram-without-check-ssl @ git+https://github.com/danissimo228/aiogram_without_check_ssl.git@b51b707f1cf74db1da5762bff6a3ac71f126c23f
youtube-search-python @ git+https://github.com/bimodinochka/ytsp.git@b7fb76222b577fc2759274a20d8f58e2d591d766
